feat(autopilotkit): extract liveness reroute helpers
